+++ go/contacts/tests.py
@@ -7,7 +7,7 @@ from django.test.client import Client
 from django.core.urlresolvers import reverse
 
 from go.base.models import User
-from go.base.tests.utils import VumiGoDjangoTestCase
+from go.apps.tests.base import DjangoGoApplicationTestCase
 from go.vumitools.contact import ContactStore
 from go.vumitools.conversation import ConversationStore
 from django.core import mail
@@ -29,7 +29,7 @@ def group_url(group_key):
     return reverse('contacts:group', kwargs={'group_key': group_key})
 
 
-class ContactsTestCase(VumiGoDjangoTestCase):
+class ContactsTestCase(DjangoGoApplicationTestCase):
 
     fixtures = ['test_user']
 
@@ -39,78 +39,15 @@ class ContactsTestCase(VumiGoDjangoTestCase):
         self.client = Client()
         self.client.login(username='username', password='password')
 
-    def setup_riak_fixtures(self):
-        self.user = User.objects.get(username='username')
-        self.contact_store = ContactStore.from_django_user(self.user)
-        self.contact_store.contacts.enable_search()
-        self.contact_store.groups.enable_search()
-
-        # We need a group
-        self.group = self.contact_store.new_group(TEST_GROUP_NAME)
-        self.group_key = self.group.key
-
-        # Also a contact
-        self.contact = self.contact_store.new_contact(
-            name=TEST_CONTACT_NAME, surname=TEST_CONTACT_SURNAME,
-            msisdn=u"+27761234567")
-        self.contact.add_to_group(self.group)
-        self.contact.save()
-        self.contact_key = self.contact.key
-
     def test_redirect_index(self):
         response = self.client.get(reverse('contacts:index'))
         self.assertRedirects(response, reverse('contacts:groups'))
 
-    def test_groups_creation(self):
-        response = self.client.post(reverse('contacts:groups'), {
-            'name': 'a new group',
-            '_new_group': '1',
-        })
-        group = newest(self.contact_store.list_groups())
-        self.assertNotEqual(group, None)
-        self.assertEqual(u'a new group', group.name)
-        self.assertRedirects(response, group_url(group.key))
-
-    def test_groups_creation_with_funny_chars(self):
-        response = self.client.post(reverse('contacts:groups'), {
-            'name': "a new group! with cüte chars's",
-            '_new_group': '1',
-        })
-        group = newest(self.contact_store.list_groups())
-        self.assertNotEqual(group, None)
-        self.assertEqual(u"a new group! with cüte chars's", group.name)
-        self.assertRedirects(response, group_url(group.key))
-
-    def test_group_contact_querying(self):
-        # test no-match
-        response = self.client.get(group_url(self.group_key), {
-            'q': 'this should not match',
-        })
-        self.assertContains(response, 'No contact match')
-
-        # test match name
-        response = self.client.get(group_url(self.group_key), {
-            'q': TEST_CONTACT_NAME,
-        })
-        self.assertContains(response, person_url(self.contact_key))
-
-    def test_group_contact_filter_by_letter(self):
-        first_letter = TEST_CONTACT_SURNAME[0]
-
-        # Assert that our name doesn't start with our "fail" case.
-        self.assertNotEqual(first_letter.lower(), 'z')
-
-        response = self.client.get(group_url(self.group_key), {'l': 'z'})
-        self.assertContains(response, 'No contact surnames start with '
-                                        'the letter')
-
-        response = self.client.get(group_url(self.group_key),
-                                   {'l': first_letter.upper()})
-        self.assertContains(response, person_url(self.contact_key))
-
-        response = self.client.get(group_url(self.group_key),
-                                   {'l': first_letter.lower()})
-        self.assertContains(response, person_url(self.contact_key))
+    def clear_groups(self, contact_key=None):
+        contact = self.contact_store.get_contact_by_key(
+            contact_key or self.contact_key)
+        contact.groups.clear()
+        contact.save()
 
     def test_contact_creation(self):
         response = self.client.post(reverse('contacts:new_person'), {
@@ -153,41 +90,6 @@ class ContactsTestCase(VumiGoDjangoTestCase):
         self.assertEqual(contact.msisdn, '112')
         self.assertEqual(set([g.key for g in contact.groups.get_all()]),
                     set([g.key for g in self.contact_store.list_groups()]))
-
-    def test_group_deletion(self):
-        # Create a contact in the group
-        response = self.client.post(reverse('contacts:new_person'), {
-            'name': 'New',
-            'surname': 'Person',
-            'msisdn': '27761234567',
-            'groups': [self.group_key],
-            })
-
-        contacts = self.contact_store.list_contacts()
-        contact = max(contacts, key=lambda c: c.created_at)
-        self.assertRedirects(response, person_url(contact.key))
-
-        # Delete the group
-        group_url = reverse('contacts:group', kwargs={
-            'group_key': self.group.key,
-        })
-        response = self.client.post(group_url, {
-                '_delete_group': True,
-            })
-        self.assertEqual(response.status_code, 302)
-        self.assertTrue(response['Location'].endswith(
-            reverse('contacts:index')))
-
-        reloaded_contacts = self.contact_store.list_contacts()
-        reloaded_contact = max(reloaded_contacts, key=lambda c: c.created_at)
-        self.assertEqual(reloaded_contact.key, contact.key)
-        self.assertEqual(reloaded_contact.groups.get_all(), [])
-
-    def clear_groups(self, contact_key=None):
-        contact = self.contact_store.get_contact_by_key(
-            contact_key or self.contact_key)
-        contact.groups.clear()
-        contact.save()
 
     def specify_columns(self, group_key=None, columns=None):
         group_url = reverse('contacts:group', kwargs={
@@ -429,7 +331,98 @@ class ContactsTestCase(VumiGoDjangoTestCase):
             'q': 'name:%s' % (self.contact.name,)
         })
 
-class SmartGroupsTestCase(VumiGoDjangoTestCase):
+class GroupsTestCase(DjangoGoApplicationTestCase):
+
+    fixtures = ['test_user']
+
+    def setUp(self):
+        super(GroupsTestCase, self).setUp()
+        self.setup_riak_fixtures()
+        self.client = Client()
+        self.client.login(username='username', password='password')
+
+    def test_groups_creation(self):
+        response = self.client.post(reverse('contacts:groups'), {
+            'name': 'a new group',
+            '_new_group': '1',
+        })
+        group = newest(self.contact_store.list_groups())
+        self.assertNotEqual(group, None)
+        self.assertEqual(u'a new group', group.name)
+        self.assertRedirects(response, group_url(group.key))
+
+    def test_groups_creation_with_funny_chars(self):
+        response = self.client.post(reverse('contacts:groups'), {
+            'name': "a new group! with cüte chars's",
+            '_new_group': '1',
+        })
+        group = newest(self.contact_store.list_groups())
+        self.assertNotEqual(group, None)
+        self.assertEqual(u"a new group! with cüte chars's", group.name)
+        self.assertRedirects(response, group_url(group.key))
+
+    def test_group_contact_querying(self):
+        # test no-match
+        response = self.client.get(group_url(self.group_key), {
+            'q': 'this should not match',
+        })
+        self.assertContains(response, 'No contact match')
+
+        # test match name
+        response = self.client.get(group_url(self.group_key), {
+            'q': TEST_CONTACT_NAME,
+        })
+        self.assertContains(response, person_url(self.contact_key))
+
+    def test_group_contact_filter_by_letter(self):
+        first_letter = TEST_CONTACT_SURNAME[0]
+
+        # Assert that our name doesn't start with our "fail" case.
+        self.assertNotEqual(first_letter.lower(), 'z')
+
+        response = self.client.get(group_url(self.group_key), {'l': 'z'})
+        self.assertContains(response, 'No contact surnames start with '
+                                        'the letter')
+
+        response = self.client.get(group_url(self.group_key),
+                                   {'l': first_letter.upper()})
+        self.assertContains(response, person_url(self.contact_key))
+
+        response = self.client.get(group_url(self.group_key),
+                                   {'l': first_letter.lower()})
+        self.assertContains(response, person_url(self.contact_key))
+
+    def test_group_deletion(self):
+        # Create a contact in the group
+        response = self.client.post(reverse('contacts:new_person'), {
+            'name': 'New',
+            'surname': 'Person',
+            'msisdn': '27761234567',
+            'groups': [self.group_key],
+            })
+
+        contacts = self.contact_store.list_contacts()
+        contact = max(contacts, key=lambda c: c.created_at)
+        self.assertRedirects(response, person_url(contact.key))
+
+        # Delete the group
+        group_url = reverse('contacts:group', kwargs={
+            'group_key': self.group.key,
+        })
+        response = self.client.post(group_url, {
+                '_delete_group': True,
+            })
+        self.assertEqual(response.status_code, 302)
+        self.assertTrue(response['Location'].endswith(
+            reverse('contacts:index')))
+
+        reloaded_contacts = self.contact_store.list_contacts()
+        reloaded_contact = max(reloaded_contacts, key=lambda c: c.created_at)
+        self.assertEqual(reloaded_contact.key, contact.key)
+        self.assertEqual(reloaded_contact.groups.get_all(), [])
+
+
+class SmartGroupsTestCase(DjangoGoApplicationTestCase):
 
     fixtures = ['test_user']
 
@@ -438,32 +431,6 @@ class SmartGroupsTestCase(VumiGoDjangoTestCase):
         self.setup_riak_fixtures()
         self.client = Client()
         self.client.login(username='username', password='password')
-
-    def setup_riak_fixtures(self):
-        self.user = User.objects.get(username='username')
-        self.contact_store = ContactStore.from_django_user(self.user)
-        self.contact_store.contacts.enable_search()
-        self.contact_store.groups.enable_search()
-
-        self.conversation_store = ConversationStore.from_django_user(self.user)
-
-        # We need a group
-        self.group = self.contact_store.new_group(TEST_GROUP_NAME)
-        self.group_key = self.group.key
-
-    def mkconversation(self, **kwargs):
-        defaults = {
-            'conversation_type': u'bulk_message',
-            'subject': u'subject',
-            'message': u'hello world'
-        }
-        defaults.update(kwargs)
-        return self.conversation_store.new_conversation(**defaults)
-
-    def mkcontact(self, name=TEST_CONTACT_NAME, surname=TEST_CONTACT_SURNAME,
-        msisdn=u'+27761234567', **kwargs):
-        return self.contact_store.new_contact(name=unicode(name),
-            surname=unicode(surname), msisdn=unicode(msisdn), **kwargs)
 
     def mksmart_group(self, query, name='a smart group'):
         response = self.client.post(reverse('contacts:groups'), {
@@ -481,12 +448,12 @@ class SmartGroupsTestCase(VumiGoDjangoTestCase):
         return contact
 
     def test_smart_groups_creation(self):
-        group = self.mksmart_group('msisdn:\+27*')
+        group = self.mksmart_group('msisdn:\+12*')
         self.assertEqual(u'a smart group', group.name)
-        self.assertEqual(u'msisdn:\+27*', group.query)
+        self.assertEqual(u'msisdn:\+12*', group.query)
 
     def test_smart_group_deletion(self):
-        group = self.mksmart_group('msisdn:\+27*')
+        group = self.mksmart_group('msisdn:\+12*')
         response = self.client.post(reverse('contacts:group', kwargs={
             'group_key': group.key,
             }), {
@@ -498,7 +465,7 @@ class SmartGroupsTestCase(VumiGoDjangoTestCase):
 
     def test_smart_group_clearing(self):
         contact = self.mkcontact()
-        group = self.mksmart_group('msisdn:\+27*')
+        group = self.mksmart_group('msisdn:\+12*')
         self.assertEqualContacts(set([contact]),
             self.contact_store.get_contacts_for_group(group))
         response = self.client.post(reverse('contacts:group', kwargs={
@@ -512,7 +479,7 @@ class SmartGroupsTestCase(VumiGoDjangoTestCase):
             self.contact_store.get_contacts_for_group(group))
 
     def test_smart_group_updating(self):
-        group = self.mksmart_group('msisdn:\+27*')
+        group = self.mksmart_group('msisdn:\+12*')
         response = self.client.post(reverse('contacts:group', kwargs={
             'group_key': group.key,
             }), {
@@ -530,7 +497,7 @@ class SmartGroupsTestCase(VumiGoDjangoTestCase):
     def test_smart_groups_no_matches_results(self):
         response = self.client.post(reverse('contacts:groups'), {
             'name': 'a smart group',
-            'query': 'msisdn:\+27*',
+            'query': 'msisdn:\+12*',
             '_new_smart_group': '1',
             })
         group = newest(self.contact_store.list_groups())
@@ -540,7 +507,7 @@ class SmartGroupsTestCase(VumiGoDjangoTestCase):
 
         self.assertRedirects(response, group_url(group.key))
         self.assertEqual(u'a smart group', group.name)
-        self.assertEqual(u'msisdn:\+27*', group.query)
+        self.assertEqual(u'msisdn:\+12*', group.query)
         self.assertEqual(
             self.contact_store.get_contacts_for_conversation(conversation),
             set([]))
@@ -556,7 +523,7 @@ class SmartGroupsTestCase(VumiGoDjangoTestCase):
     def test_smart_groups_with_matches_results(self):
         response = self.client.post(reverse('contacts:groups'), {
             'name': 'a smart group',
-            'query': 'msisdn:\+27*',
+            'query': 'msisdn:\+12*',
             '_new_smart_group': '1',
             })
 
@@ -568,7 +535,7 @@ class SmartGroupsTestCase(VumiGoDjangoTestCase):
 
         self.assertRedirects(response, group_url(group.key))
         self.assertEqual(u'a smart group', group.name)
-        self.assertEqual(u'msisdn:\+27*', group.query)
+        self.assertEqual(u'msisdn:\+12*', group.query)
         self.assertEqual(
             self.contact_store.get_static_contacts_for_group(group), [])
         self.assertEqualContacts(
